requirements.txt
requests==2.7.0
instabot==0.4.5
python-dotenv==0.10.3